Fixed all incorrect names
