+++ Recruitment/RecruitmentAgency/RecruitmentAgencyServer/RecruitmentAgencyServer.csproj
@@ -5,7 +5,7 @@
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<GenerateDocumentationFile>True</GenerateDocumentationFile>
-		<StartupObject>ApplicationRecruitmentAgency.Server</StartupObject>
+		<StartupObject>RecruitmentAgencyServer.Server</StartupObject>
 	</PropertyGroup>
 
 	<ItemGroup>
