+++ Recruitment/RecruitmentAgency/RecruitmentAgencyServer/Repository/IApplicationsServerRepository.cs
